Dir: SearchHeader 경로 변경, 스타일 코드 분리 #81
